// src/models/playlist.model.ts
import mongoose, { Schema, Document } from "mongoose";

export interface IPlaylist extends Document {
    name: string;
    description: string;
    videos: mongoose.Types.ObjectId[];
    owner: mongoose.Types.ObjectId;
    channel: mongoose.Types.ObjectId; // one to whom 'subscriber' is subscribing
    notifications: boolean;
    isPrivate: boolean;
    createdAt: Date;
    updatedAt: Date;
}

const playlistSchema = new Schema<IPlaylist>({
    name: {
        type: String,
        required: true
    },
    description: {
        type: String,
        required: true
    },
    isPrivate: {
        type: Boolean,
        default: false
    },
    videos: [
        {
            type: Schema.Types.ObjectId,
            ref: "Video"
        }
    ],
    owner: {
        type: Schema.Types.ObjectId,
        ref: "User"
    }
}, { timestamps: true })

export const Playlist = mongoose.model<IPlaylist>("Playlist", playlistSchema);